+++ tests/unit/pipelex/core/concepts/test_concept_compact_memory.py
